Moved tables in kk_token.
Moved all but the TOKENS table into the objects which use them. Perhaps this
hurts readability, perhaps not. It does mean that you have to actually scan
through the files to see which tables are where, and makes changing the grammar
slightly less straightforward, but I think that encapsulating the tables in
their respective objects will hopefully lead to more straightforward code with
respect to the objects.